+++ app/Http/Controllers/Admin/JobCategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\JobCategory;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -16,14 +17,14 @@ class JobCategoryController extends Controller
      * Display a listing of the resource.
      */
     use Searchable;
-    public function index() : View
+    public function index(): View
     {
         $query = JobCategory::query();
 
         $this->search($query, ['name']);
 
-        $jobCategories = $query->withCount(['jobs' => function($query){
-            $query->where('status','active')->where('deadline', '>=', date('Y-m-d'));
+        $jobCategories = $query->withCount(['jobs' => function ($query) {
+            $query->where('status', 'active')->where('deadline', '>=', date('Y-m-d'));
         }])->paginate(10);
         return view('admin.job.job-category.index', compact(
             'jobCategories'
@@ -33,7 +34,7 @@ class JobCategoryController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create() : View
+    public function create(): View
     {
         return view('admin.job.job-category.create');
     }
@@ -61,13 +62,12 @@ class JobCategoryController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id) : View
+    public function edit(string $id): View
     {
         $jobCategory = JobCategory::findOrFail($id);
         return view('admin.job.job-category.edit', compact(
             'jobCategory'
         ));
-
     }
 
     /**
@@ -81,7 +81,7 @@ class JobCategoryController extends Controller
         ]);
 
         $jobCategory = JobCategory::findOrFail($id);
-        if($request->filled('icon')) $jobCategory->icon = $request->icon;
+        if ($request->filled('icon')) $jobCategory->icon = $request->icon;
         $jobCategory->name = $request->name;
         $jobCategory->save();
 
@@ -95,6 +95,10 @@ class JobCategoryController extends Controller
      */
     public function destroy(string $id)
     {
+        $jobExist = Job::where('job_category_id', $id)->exists();
+        if ($jobExist) {
+            return response(['message' => 'error'], 500);
+        }
         try {
             JobCategory::findOrFail($id)->delete();
             Notify::DeleteNotify();
@@ -105,7 +109,8 @@ class JobCategoryController extends Controller
         }
     }
 
-    public function changeStatus(string $id) : Response {
+    public function changeStatus(string $id): Response
+    {
         $jobCategory = JobCategory::FindOrFail($id);
         $jobCategory->featured = $jobCategory->featured == 0 ? 1 : 0;
         $jobCategory->save();
