+++ app/Http/Controllers/Admin/SkillController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\JobSkill;
 use App\Models\Skill;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -86,6 +87,10 @@ class SkillController extends Controller
      */
     public function destroy(string $id)
     {
+        $jobExist = JobSkill::where('skill_id', $id)->exists();
+        if ($jobExist) {
+            return response(['message' => 'error'], 500);
+        }
         try {
             $skill = Skill::findOrFail($id)->delete();
             Notify::DeleteNotify();
